refactor: mv params parse and valid part to pkg package that client and daemon can both use
