+++ pkg/opts/intel_rdt.go
@@ -1,0 +1,9 @@
+package opts
+
+// ParseIntelRdt parses inter-rdt params of container
+func ParseIntelRdt(intelRdtL3Cbm string) (string, error) {
+	// FIXME: add Intel RDT L3 Cbm validation
+	return intelRdtL3Cbm, nil
+}
+
+// TODO: ValidateInterRdt
