+++ pkg/opts/portbindings_test.go
@@ -1,0 +1,54 @@
+package opts
+
+import (
+	"reflect"
+	"testing"
+
+	"github.com/alibaba/pouch/apis/types"
+)
+
+func TestParsePortBinding(t *testing.T) {
+	type args struct {
+		ports []string
+	}
+	tests := []struct {
+		name    string
+		args    args
+		want    types.PortMap
+		wantErr bool
+	}{
+	// TODO: Add test cases.
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := ParsePortBinding(tt.args.ports)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("ParsePortBinding() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("ParsePortBinding() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func TestVerifyPortBinding(t *testing.T) {
+	type args struct {
+		portBindings types.PortMap
+	}
+	tests := []struct {
+		name    string
+		args    args
+		wantErr bool
+	}{
+	// TODO: Add test cases.
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if err := ValidatePortBinding(tt.args.portBindings); (err != nil) != tt.wantErr {
+				t.Errorf("VerifyPortBinding() error = %v, wantErr %v", err, tt.wantErr)
+			}
+		})
+	}
+}
