+++ pkg/opts/ports_test.go
@@ -1,0 +1,33 @@
+package opts
+
+import (
+	"reflect"
+	"testing"
+)
+
+func TestParseExposedPorts(t *testing.T) {
+	type args struct {
+		portList []string
+		expose   []string
+	}
+	tests := []struct {
+		name    string
+		args    args
+		want    map[string]interface{}
+		wantErr bool
+	}{
+	// TODO: Add test cases.
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := ParseExposedPorts(tt.args.portList, tt.args.expose)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("ParseExposedPorts() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("ParseExposedPorts() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
